+++ src/core/settings.ts
@@ -3,8 +3,8 @@ export default class SETTINGS {
 	static init(moduleName: string) {
 		this._MOD_NAME = moduleName;
 	}
-	static register<T>(key: string, config: ClientSettings.PartialSetting<T>) { game.settings.register(SETTINGS._MOD_NAME, key, config); }
-	static registerMenu(key: string, config: ClientSettings.PartialMenuSetting) { game.settings.registerMenu(SETTINGS._MOD_NAME, key, config); }
+	static register<T>(key: string, config: ClientSettings.PartialSettingConfig<T>) { game.settings.register(SETTINGS._MOD_NAME, key, config); }
+	static registerMenu(key: string, config: ClientSettings.PartialSettingSubmenuConfig) { game.settings.registerMenu(SETTINGS._MOD_NAME, key, config); }
 	static get<T>(key: string): T { return <T>game.settings.get(SETTINGS._MOD_NAME, key); }
 	static async set<T>(key: string, value: T): Promise<T> { return await game.settings.set(SETTINGS._MOD_NAME, key, value); }
 	static default<T>(key: string): T { return <T>game.settings.settings.get(SETTINGS._MOD_NAME + '.' + key).default; }